Day 24
It is sufficient to use three hailstones to figure out
the required values (three points of intersection are
sufficient to describe the line, and we know exactly
one solution line exists).

Add a Rational type, so we don't incur floating point
errors, as everything becomes integer arithmetic. The
implicits are cobbled together from BigDecimal and mean
we can write sensible looking code with minimal
implementation.