2015 Day 24
Wasted a lot of time trying to do a clever solution which
works out the viable groups using a tree. It was far too slow.

In the end (taking advice from the subreddit), find all the combinations
up to 1/3 of the package count (the minimum group must be at most this large),
and find the smallest one with the smallest product.