2016 Day 22
Insight required to detect the naive infeasibility of the solution,
and the cost function (though having thought about it the cost
makes sense).

Add caching to Dijkstra and update the 2015 problem accordingly.

https://codepen.io/anon/pen/BQEZzK/ used to visualize.

Use the visualization to make the leap that the cost function
has you bump up against the wall, so you need to navigate round
it manually.